improve _putchar function in utils.c
improve _putchar function to minimise the number of write system
calls made. This implementation ensures that the characters in the buffer are
correctly flushed to the standard output, either when the buffer is about to be full
or if there are remaining characters when the printing is finished.
Additionally, an update was made to the main.h file with BUFFSIZE macro.